requirements.txt
# --- Force CPU-only Torch ---
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
torchvision==0.17.1+cpu

# --- Member 3: Backend & Database ---
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
pydantic==2.5.2
httpx==0.25.2

# --- Member 2: RAG Pipeline ---
langchain>=0.1.11
langchain-community>=0.0.28
langchain-core>=0.1.33
langchain-chroma==0.1.0
langchain-groq==0.1.0
langchain-huggingface==0.0.1
chromadb==0.4.22
pypdf==3.17.4
# ADD THIS LINE:
cryptography>=42.0.0 
python-dotenv==1.0.0
sentence-transformers>=3.0.1 
rank-bm25>=0.2.2

# --- Member 5 & Week 10 Testing ---
pytest>=7.4.0
scikit-learn==1.3.2
numpy>=1.26.4
pandas==2.1.3
ragas>=0.2.0
datasets>=2.16.1
openai>=1.0.0
anthropic>=0.18.0